Moved attendees to domain
