+++ src/Domain/Attendees/index.tsx
@@ -1,0 +1,5 @@
+export interface IAttendees {
+    Councillors: string,
+    HackneyStaff: string,
+    NumberOfAttendees: number
+}
